flask/requirements.txt
flask==2.3.2
cassandra-driver==3.29.0
Werkzeug==2.3.7
